Minor reorganization.
